+++ ChangeLog.md
@@ -1,3 +1,6 @@
 # V 0.0.7
  - Enhanced face orientaion/position estimation
  - Bugfixes : process function do not do mask plotting on faces from now on
+# V 0.0.8
+ - Repared face orientation estimation and recentered the angles
+ - Added eye position estimation and an exemple of how to build eye orientation in space